requirements.txt
# Core Frameworks
streamlit
langchain
langgraph
langchain_google_genai
python-dotenv

# Data & Analysis
yfinance
pandas
pandas-ta

# Plotting
plotly

# NLP & ML
transformers
torch
scikit-learn

